+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/DefaultNativeConnectionProvider.java
@@ -51,6 +51,7 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
         String host = nativeConfig.getHost();
         int port = nativeConfig.getPort();
         boolean remoteRouting = nativeConfig.getRemoteRouting();
+        String consistencySerial = nativeConfig.getConsistencySerial();
         Security.CqlSecurity cqlSecurity = cqlSecuritySupplier.get();
         boolean authEnabled = cqlSecurity.getCredentials().isEnabled();
         boolean tlsEnabled = cqlSecurity.getTls().isEnabled();
@@ -72,6 +73,7 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
                 .withLocalhost(host)
                 .withPort(port)
                 .withRemoteRouting(remoteRouting)
+                .withConsistencySerial(consistencySerial)
                 .withAuthProvider(authProvider)
                 .withSslEngineFactory(sslEngineFactory)
                 .withMetricsEnabled(config.getStatistics().isEnabled())
@@ -142,6 +144,12 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
     public final boolean getRemoteRouting()
     {
         return myLocalNativeConnectionProvider.getRemoteRouting();
+    }
+
+    @Override
+    public final String getSerialConsistency()
+    {
+        return myLocalNativeConnectionProvider.getSerialConsistency();
     }
 
     @Override