+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestDefaultRepairConfigurationProvider.java
@@ -99,6 +99,11 @@ public class TestDefaultRepairConfigurationProvider
             {
                 return true;
             }
+
+            @Override
+            public String getSerialConsistency(){
+                return "DEFAULT";
+            }
         };
 
         when(session.getMetadata()).thenReturn(metadata);